+++ lib/article_json/export/facebook_instant_article/elements/heading.rb
@@ -1,0 +1,11 @@
+module ArticleJSON
+  module Export
+    module FacebookInstantArticle
+      module Elements
+        class Heading < Base
+          include ArticleJSON::Export::Common::HTML::Elements::Heading
+        end
+      end
+    end
+  end
+end