+++ lib/article_json/export/facebook_instant_article/elements/list.rb
@@ -1,0 +1,11 @@
+module ArticleJSON
+  module Export
+    module FacebookInstantArticle
+      module Elements
+        class List < Base
+          include ArticleJSON::Export::Common::HTML::Elements::List
+        end
+      end
+    end
+  end
+end