+++ lib/article_json/export/facebook_instant_article/elements/text_box.rb
@@ -1,0 +1,11 @@
+module ArticleJSON
+  module Export
+    module FacebookInstantArticle
+      module Elements
+        class TextBox < Base
+          include ArticleJSON::Export::Common::HTML::Elements::TextBox
+        end
+      end
+    end
+  end
+end